requirements.txt
numpy
scipy
Pillow